# api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Discodeit API",
    "description": "Sprint Mission",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/messages": {
      "get": {
        "tags": [
          "Message"
        ],
        "summary": "Channel의 Message 목록 조회",
        "operationId": "findAllByChannelId",
        "parameters": [
          {
            "name": "channelId",
            "in": "query",
            "description": "조회할 Channel ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Message 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Message"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Message"
        ],
        "summary": "Message 생성",
        "operationId": "create_2",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "messageCreateRequest"
                ],
                "type": "object",
                "properties": {
                  "messageCreateRequest": {
                    "$ref": "#/components/schemas/MessageCreateRequest"
                  },
                  "attachments": {
                    "type": "array",
                    "description": "Message 첨부 파일들",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message가 성공적으로 생성됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Channel 또는 User를 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "Channel | Author with id {channelId | authorId} not found"
              }
            }
          }
        }
      }
    },
    "/api/channels/public": {
      "post": {
        "tags": [
          "Channel"
        ],
        "summary": "Public Channel 생성",
        "operationId": "create_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicChannelCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Public Channel이 성공적으로 생성됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          }
        }
      }
    },
    "/api/channels/private": {
      "post": {
        "tags": [
          "Channel"
        ],
        "summary": "Private Channel 생성",
        "operationId": "create_4",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrivateChannelCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Private Channel이 성공적으로 생성됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/messages/{messageId}": {
      "delete": {
        "tags": [
          "Message"
        ],
        "summary": "Message 삭제",
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "description": "삭제할 Message ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Message가 성공적으로 삭제됨"
          },
          "404": {
            "description": "Message를 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "Message with id {messageId} not found"
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Message"
        ],
        "summary": "Message 내용 수정",
        "operationId": "update_2",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "description": "수정할 Message ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Message가 성공적으로 수정됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "404": {
            "description": "Message를 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "Message with id {messageId} not found"
              }
            }
          }
        }
      }
    },
    "/api/channels/{channelId}": {
      "delete": {
        "tags": [
          "Channel"
        ],
        "summary": "Channel 삭제",
        "operationId": "delete_2",
        "parameters": [
          {
            "name": "channelId",
            "in": "path",
            "description": "삭제할 Channel ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Channel을 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "Channel with id {channelId} not found"
              }
            }
          },
          "204": {
            "description": "Channel이 성공적으로 삭제됨"
          }
        }
      },
      "patch": {
        "tags": [
          "Channel"
        ],
        "summary": "Channel 정보 수정",
        "operationId": "update_3",
        "parameters": [
          {
            "name": "channelId",
            "in": "path",
            "description": "수정할 channel ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicChannelUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "404": {
            "description": "Channel을 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "Private channel cannot be updated"
              }
            }
          },
          "400": {
            "description": "Private Channel은 수정할 수 없음",
            "content": {
              "*/*": {
                "example": "Channel with id {channel} not found"
              }
            }
          },
          "200": {
            "description": "Channel 정보가 성공적으로 수정됨",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Channel"
                }
              }
            }
          }
        }
      }
    },
    "/api/channels": {
      "get": {
        "tags": [
          "Channel"
        ],
        "summary": "User가 참여 중인 Channel 목록 조회",
        "operationId": "findAll_1",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "description": "조회할 User ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Channel 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChannelDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/binaryContents": {
      "get": {
        "tags": [
          "BinaryContent"
        ],
        "summary": "여러 첨부 파일 조회",
        "operationId": "findAllByIdIn",
        "parameters": [
          {
            "name": "binaryContentIds",
            "in": "query",
            "description": "조회할 첨부 파일 ID 목록",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "첨부 파일 목록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BinaryContent"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/binaryContents/{binaryContentId}": {
      "get": {
        "tags": [
          "BinaryContent"
        ],
        "summary": "첨부 파일 조회",
        "operationId": "find",
        "parameters": [
          {
            "name": "binaryContentId",
            "in": "path",
            "description": "조회할 첨부 파일 ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "첨부 파일 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BinaryContent"
                }
              }
            }
          },
          "404": {
            "description": "첨부 파일을 찾을 수 없음",
            "content": {
              "*/*": {
                "example": "BinaryContent with id {binaryContentId} not found"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "MessageCreateRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          },
          "channelId": {
            "type": "string",
            "format": "uuid"
          },
          "authorId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "description": "Message 생성 정보"
      },
      "Message": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "content": {
            "type": "string"
          },
          "channelId": {
            "type": "string",
            "format": "uuid"
          },
          "authorId": {
            "type": "string",
            "format": "uuid"
          },
          "attachmentIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "PublicChannelCreateRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "Channel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          }
        }
      },
      "PrivateChannelCreateRequest": {
        "type": "object",
        "properties": {
          "participantIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
...
      "MessageUpdateRequest": {
        "type": "object",
        "properties": {
          "newContent": {
            "type": "string"
          }
        }
      },
      "PublicChannelUpdateRequest": {
        "type": "object",
        "properties": {
          "newName": {
            "type": "string"
          },
          "newDescription": {
            "type": "string"
          }
        },
        "description": "수정할 channel 정보"
      },
...
      "ChannelDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PRIVATE"
            ]
          },
          "lastMessageAt": {
            "type": "string",
            "format": "date-time"
          },
          "participantIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        }
      },
      "BinaryContent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "bytes": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "byte"
            }
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    }
  }
}